+++ lib/router/routes.g.dart
@@ -8,39 +8,15 @@ part of 'routes.dart';
 
 List<RouteBase> get $appRoutes => [
       $homeRoute,
+      $loginRoute,
+      $verifyEmailRoute,
+      $waitingForApprovalRoute,
+      $settingsRoute,
     ];
 
 RouteBase get $homeRoute => GoRouteData.$route(
       path: '/',
       factory: $HomeRouteExtension._fromState,
-      routes: [
-        GoRouteData.$route(
-          path: 'login',
-          factory: $LoginRouteExtension._fromState,
-        ),
-        GoRouteData.$route(
-          path: 'sign-up',
-          factory: $SignUpRouteExtension._fromState,
-        ),
-        GoRouteData.$route(
-          path: 'users',
-          factory: $UsersRouteExtension._fromState,
-        ),
-        GoRouteData.$route(
-          path: 'events',
-          factory: $EventsRouteExtension._fromState,
-          routes: [
-            GoRouteData.$route(
-              path: 'create',
-              factory: $EventCreateRouteExtension._fromState,
-            ),
-          ],
-        ),
-        GoRouteData.$route(
-          path: 'settings',
-          factory: $SettingsRouteExtension._fromState,
-        ),
-      ],
     );
 
 extension $HomeRouteExtension on HomeRoute {
@@ -59,6 +35,17 @@ extension $HomeRouteExtension on HomeRoute {
 
   void replace(BuildContext context) => context.replace(location);
 }
+
+RouteBase get $loginRoute => GoRouteData.$route(
+      path: '/login',
+      factory: $LoginRouteExtension._fromState,
+      routes: [
+        GoRouteData.$route(
+          path: 'sign-up',
+          factory: $SignUpRouteExtension._fromState,
+        ),
+      ],
+    );
 
 extension $LoginRouteExtension on LoginRoute {
   static LoginRoute _fromState(GoRouterState state) => LoginRoute(
@@ -87,76 +74,7 @@ extension $SignUpRouteExtension on SignUpRoute {
   static SignUpRoute _fromState(GoRouterState state) => const SignUpRoute();
 
   String get location => GoRouteData.$location(
-        '/sign-up',
-      );
-
-  void go(BuildContext context) => context.go(location);
-
-  Future<T?> push<T>(BuildContext context) => context.push<T>(location);
-
-  void pushReplacement(BuildContext context) =>
-      context.pushReplacement(location);
-
-  void replace(BuildContext context) => context.replace(location);
-}
-
-extension $UsersRouteExtension on UsersRoute {
-  static UsersRoute _fromState(GoRouterState state) => const UsersRoute();
-
-  String get location => GoRouteData.$location(
-        '/users',
-      );
-
-  void go(BuildContext context) => context.go(location);
-
-  Future<T?> push<T>(BuildContext context) => context.push<T>(location);
-
-  void pushReplacement(BuildContext context) =>
-      context.pushReplacement(location);
-
-  void replace(BuildContext context) => context.replace(location);
-}
-
-extension $EventsRouteExtension on EventsRoute {
-  static EventsRoute _fromState(GoRouterState state) => const EventsRoute();
-
-  String get location => GoRouteData.$location(
-        '/events',
-      );
-
-  void go(BuildContext context) => context.go(location);
-
-  Future<T?> push<T>(BuildContext context) => context.push<T>(location);
-
-  void pushReplacement(BuildContext context) =>
-      context.pushReplacement(location);
-
-  void replace(BuildContext context) => context.replace(location);
-}
-
-extension $EventCreateRouteExtension on EventCreateRoute {
-  static EventCreateRoute _fromState(GoRouterState state) =>
-      const EventCreateRoute();
-
-  String get location => GoRouteData.$location(
-        '/events/create',
-      );
-
-  void go(BuildContext context) => context.go(location);
-
-  Future<T?> push<T>(BuildContext context) => context.push<T>(location);
-
-  void pushReplacement(BuildContext context) =>
-      context.pushReplacement(location);
-
-  void replace(BuildContext context) => context.replace(location);
-}
-
-extension $SettingsRouteExtension on SettingsRoute {
-  static SettingsRoute _fromState(GoRouterState state) => const SettingsRoute();
-
-  String get location => GoRouteData.$location(
-        '/settings',
+        '/login/sign-up',
       );
 
   void go(BuildContext context) => context.go(location);
@@ -176,4 +94,72 @@ T? _$convertMapValue<T>(
 ) {
   final value = map[key];
   return value == null ? null : converter(value);
+}
+
+RouteBase get $verifyEmailRoute => GoRouteData.$route(
+      path: '/verify-email',
+      factory: $VerifyEmailRouteExtension._fromState,
+    );
+
+extension $VerifyEmailRouteExtension on VerifyEmailRoute {
+  static VerifyEmailRoute _fromState(GoRouterState state) =>
+      const VerifyEmailRoute();
+
+  String get location => GoRouteData.$location(
+        '/verify-email',
+      );
+
+  void go(BuildContext context) => context.go(location);
+
+  Future<T?> push<T>(BuildContext context) => context.push<T>(location);
+
+  void pushReplacement(BuildContext context) =>
+      context.pushReplacement(location);
+
+  void replace(BuildContext context) => context.replace(location);
+}
+
+RouteBase get $waitingForApprovalRoute => GoRouteData.$route(
+      path: '/waiting-for-approval',
+      factory: $WaitingForApprovalRouteExtension._fromState,
+    );
+
+extension $WaitingForApprovalRouteExtension on WaitingForApprovalRoute {
+  static WaitingForApprovalRoute _fromState(GoRouterState state) =>
+      const WaitingForApprovalRoute();
+
+  String get location => GoRouteData.$location(
+        '/waiting-for-approval',
+      );
+
+  void go(BuildContext context) => context.go(location);
+
+  Future<T?> push<T>(BuildContext context) => context.push<T>(location);
+
+  void pushReplacement(BuildContext context) =>
+      context.pushReplacement(location);
+
+  void replace(BuildContext context) => context.replace(location);
+}
+
+RouteBase get $settingsRoute => GoRouteData.$route(
+      path: '/settings',
+      factory: $SettingsRouteExtension._fromState,
+    );
+
+extension $SettingsRouteExtension on SettingsRoute {
+  static SettingsRoute _fromState(GoRouterState state) => const SettingsRoute();
+
+  String get location => GoRouteData.$location(
+        '/settings',
+      );
+
+  void go(BuildContext context) => context.go(location);
+
+  Future<T?> push<T>(BuildContext context) => context.push<T>(location);
+
+  void pushReplacement(BuildContext context) =>
+      context.pushReplacement(location);
+
+  void replace(BuildContext context) => context.replace(location);
 }