+++ lib/screens/waiting_for_approval/waiting_for_approval_screen.dart
@@ -1,0 +1,10 @@
+import 'package:flutter/material.dart';
+
+class WaitingForApprovalScreen extends StatelessWidget {
+  const WaitingForApprovalScreen({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return const Placeholder();
+  }
+}